referencing issue #56
